javascriptのテストを別ディレクトリに切り出し
